# openapi/nexusleo.openapi.json
{
  "components": {
    "schemas": {
...
      "CaseCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CaseCreate",
        "type": "object"
      },
      "CaseExportAuditEvent": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_refs_json": {
            "additionalProperties": true,
            "title": "Input Refs Json",
            "type": "object"
          },
          "output_refs_json": {
            "additionalProperties": true,
            "title": "Output Refs Json",
            "type": "object"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          },
          "tool_version": {
            "title": "Tool Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "actor",
          "action",
          "tool",
          "tool_version",
          "input_refs_json",
          "output_refs_json",
          "created_at"
        ],
        "title": "CaseExportAuditEvent",
        "type": "object"
      },
      "CaseExportBundle": {
        "properties": {
          "audit_events": {
            "items": {
              "$ref": "#/components/schemas/CaseExportAuditEvent"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "case": {
            "$ref": "#/components/schemas/CaseExportCase"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimOut"
            },
            "title": "Claims",
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/CaseExportDocument"
            },
            "title": "Documents",
            "type": "array"
          }
        },
        "required": [
          "case",
          "documents",
          "claims",
          "audit_events"
        ],
        "title": "CaseExportBundle",
        "type": "object"
      },
      "CaseExportCase": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "CaseExportCase",
        "type": "object"
      },
      "CaseExportDocument": {
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "type",
          "sha256",
          "created_at",
          "metadata_json"
        ],
        "title": "CaseExportDocument",
        "type": "object"
      },
      "CaseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "created_at"
        ],
        "title": "CaseOut",
        "type": "object"
      },
      "ClaimOut": {
        "properties": {
          "audit_event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Audit Event Ids",
            "type": "array"
          },
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "claim_type": {
            "title": "Claim Type",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "object_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Entity Id"
          },
          "predicate": {
            "title": "Predicate",
            "type": "string"
          },
          "subject_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Entity Id"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "claim_type",
          "predicate",
          "text",
          "created_at",
          "evidence",
          "audit_event_ids"
        ],
        "title": "ClaimOut",
        "type": "object"
      },
      "ConfidenceOut": {
        "properties": {
          "factors_json": {
            "additionalProperties": true,
            "title": "Factors Json",
            "type": "object"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "rationale_text": {
            "title": "Rationale Text",
            "type": "string"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          }
        },
        "required": [
          "level",
          "rubric_version",
          "factors_json",
          "rationale_text"
        ],
        "title": "ConfidenceOut",
        "type": "object"
      },
...
      "DocumentCreate": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "default": "TRANSCRIPT",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "DocumentCreate",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "type",
          "sha256",
          "created_at",
          "metadata_json"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "EvidenceOut": {
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "evidence_link_id": {
            "format": "uuid",
            "title": "Evidence Link Id",
            "type": "string"
          },
          "mention": {
            "$ref": "#/components/schemas/MentionOut"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "evidence_link_id",
          "weight",
          "mention",
          "document_id"
        ],
        "title": "EvidenceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCaseRequest": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "RunCaseRequest",
        "type": "object"
      },
      "RunCaseResponse": {
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ClaimOut"
            },
            "title": "Claims",
            "type": "array"
          },
          "ingest_result": {
            "$ref": "#/components/schemas/IngestResult"
          }
        },
        "required": [
          "ingest_result",
          "claims"
        ],
        "title": "RunCaseResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NexusLEO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cases": {
      "post": {
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/claims": {
      "get": {
        "operationId": "list_claims_cases__case_id__claims_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "minimum": 0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClaimOut"
                  },
                  "title": "Response List Claims Cases  Case Id  Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Claims",
        "tags": [
          "claims"
        ]
      }
    },
    "/cases/{case_id}/documents": {
      "post": {
        "operationId": "create_document_cases__case_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate",
                "examples": {
                  "transcript": {
                    "summary": "Transcript",
                    "value": {
                      "metadata": {
                        "source": "demo"
                      },
                      "text": "A: John Smith met Jane Doe in Baltimore.\nB: John Smith met Jane Doe in Baltimore.",
                      "type": "TRANSCRIPT"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/cases/{case_id}/export": {
      "get": {
        "operationId": "export_case_cases__case_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "minimum": 0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "audit_events": [],
                  "case": {
                    "created_at": "2025-01-01T00:00:00Z",
                    "id": "00000000-0000-0000-0000-000000000000",
                    "name": "Demo Case"
                  },
                  "claims": [],
                  "documents": []
                },
                "schema": {
                  "$ref": "#/components/schemas/CaseExportBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Case",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/cases/{case_id}/run": {
      "post": {
        "operationId": "run_case_cases__case_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCaseRequest",
                "examples": {
                  "run": {
                    "summary": "Run case ingest",
                    "value": {
                      "document_id": "00000000-0000-0000-0000-000000000000"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Case",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}